Break out `util.urls`, `util.files`, and `util.gcs`
Just more refactoring.
