+++ bionic/flow.py
@@ -41,13 +41,13 @@ from .provider import (
 from .deriver import EntityDeriver, entity_is_internal
 from .descriptors.parsing import entity_dnode_from_descriptor
 from . import decorators, decoration
+from .filecopier import FileCopier
 from .oneline import oneline
+from .util.gcs import copy_to_gcs
 from .util.misc import (
     group_pairs,
     check_exactly_one_present,
     check_at_most_one_present,
-    copy_to_gcs,
-    FileCopier,
 )
 
 DEFAULT_PROTOCOL = protos.CombinedProtocol(
